Correctly calculate waveform width by using Double (OT-770) (#909)

+++ jvm/markerapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/markerapp/app/viewmodel/VerseMarkerViewModel.kt
@@ -234,8 +234,8 @@ class VerseMarkerViewModel : ViewModel(), IMarkerViewModel {
         }
 
         val samplesPerScreenWidth = sampleRate * secondsOnScreen
-        val samplesPerPixel = samplesPerScreenWidth / width
+        val samplesPerPixel = samplesPerScreenWidth / width.toDouble()
         val pixelsInDuration = waveformAudioPlayerProperty.get().getDurationInFrames() / samplesPerPixel
-        return pixelsInDuration.toDouble()
+        return pixelsInDuration
     }
 }
